+++ lib/appsignal/capistrano.rb
@@ -1,6 +1,7 @@
 require 'appsignal'
+require 'capistrano/version'
 
-if defined?(Capistrano::VERSION)
+if defined?(Capistrano::VERSION) && Gem::Version.new(Capistrano::VERSION) >= Gem::Version.new(3)
   # Capistrano 3+
   load File.expand_path('../integrations/capistrano/appsignal.cap', __FILE__)
 else