+++ lib/appsignal/integrations/sinatra.rb
@@ -14,6 +14,5 @@ Appsignal.start_logger(app_settings.root)
 Appsignal.start
 
 if Appsignal.active?
-  ::Sinatra::Application.use(Appsignal::Rack::Listener)
   ::Sinatra::Application.use(Appsignal::Rack::SinatraInstrumentation)
 end