+++ lib/appsignal/rack/rails_instrumentation.rb
@@ -2,9 +2,9 @@ require 'rack'
 
 module Appsignal
   module Rack
-    class SinatraInstrumentation
+    class RailsInstrumentation
       def initialize(app, options = {})
-        Appsignal.logger.debug 'Initializing Appsignal::Rack::SinatraInstrumentation'
+        Appsignal.logger.debug 'Initializing Appsignal::Rack::RailsInstrumentation'
         @app, @options = app, options
       end
 
@@ -17,30 +17,22 @@ module Appsignal
       end
 
       def call_with_appsignal_monitoring(env)
-        if @options[:params_method]
-          env[:params_method] = @options[:params_method]
-        end
-        request = @options.fetch(:request_class, Sinatra::Request).new(env)
+        request = ActionDispatch::Request.new(env)
         transaction = Appsignal::Transaction.create(
-          SecureRandom.uuid,
+          env['action_dispatch.request_id'],
           Appsignal::Transaction::HTTP_REQUEST,
           request
         )
         begin
-          ActiveSupport::Notifications.instrument('process_action.sinatra') do
-            @app.call(env)
-          end
+          @app.call(env)
         rescue => error
           transaction.set_error(error)
           raise error
         ensure
-          # In production newer versions of Sinatra don't raise errors, but store
-          # them in the sinatra.error env var.
-          transaction.set_error(env['sinatra.error']) if env['sinatra.error']
-          transaction.set_action(env['sinatra.route'])
+          transaction.set_http_or_background_action
+          transaction.set_http_or_background_queue_start
           transaction.set_metadata('path', request.path)
           transaction.set_metadata('method', request.request_method)
-          transaction.set_http_or_background_queue_start
           Appsignal::Transaction.complete_current!
         end
       end