+++ spec/lib/appsignal/params_sanitizer_spec.rb
@@ -6,39 +6,54 @@ class ErrorOnInspect
   end
 end
 
+class ClassWithInspect
+  def inspect
+    "#<ClassWithInspect foo=\"bar\"/>"
+  end
+end
+
 describe Appsignal::ParamsSanitizer do
   let(:klass) { Appsignal::ParamsSanitizer }
   let(:file) { uploaded_file }
   let(:params) do
     {
-      :text => 'string',
-      :file => file,
-      :hash => {
-        :nested_text => 'string',
+      :text       => 'string',
+      :file       => file,
+      :float      => 0.0,
+      :bool_true  => true,
+      :bool_false => false,
+      :int        => 1,
+      :hash       => {
+        :nested_text  => 'string',
         :nested_array => [
           'something',
           'else',
           file,
           {
-            :key => 'value',
+            :key  => 'value',
             :file => file,
           },
-          ErrorOnInspect.new
+          ErrorOnInspect.new,
+          ClassWithInspect.new
         ]
       }
     }
   end
   let(:sanitized_params) { klass.sanitize(params) }
-  let(:scrubbed_params) { klass.scrub(params) }
+  let(:scrubbed_params)  { klass.scrub(params) }
 
   describe ".sanitize!" do
     subject { params }
     before { klass.sanitize!(subject) }
 
     it { should be_instance_of Hash }
-    its([:text]) { should == 'string' }
-    its([:file]) { should be_instance_of String }
-    its([:file]) { should include '::UploadedFile' }
+    its([:text])        { should == 'string' }
+    its([:file])        { should be_instance_of String }
+    its([:file])        { should include '::UploadedFile' }
+    its([:float])       { should == 0.0 }
+    its([:int])         { should == 1 }
+    its([:bool_true])   { should == 'true' }
+    its([:bool_false])  { should == 'false' }
 
     context "hash" do
       subject { params[:hash] }
@@ -54,7 +69,8 @@ describe Appsignal::ParamsSanitizer do
         its([1]) { should == 'else' }
         its([2]) { should be_instance_of String }
         its([2]) { should include '::UploadedFile' }
-        its([4]) { should == '#<ErrorOnInspect/>' }
+        its([4]) { should == '#<ErrorOnInspect>' }
+        its([5]) { should == '#<ClassWithInspect>' }
 
         context "nested hash" do
           subject { params[:hash][:nested_array][3] }